Extract LineWidth and TabSize into line_width.rs
