+++ crates/ruff/src/message/text.rs
@@ -14,7 +14,7 @@ use crate::fs::relativize_path;
 use crate::message::diff::Diff;
 use crate::message::{Emitter, EmitterContext, Message};
 use crate::registry::AsRule;
-use crate::settings::options::{LineWidth, TabSize};
+use crate::settings::line_width::{LineWidth, TabSize};
 
 bitflags! {
     #[derive(Default)]
