+++ crates/ruff/src/rules/flake8_simplify/rules/ast_if.rs
@@ -15,7 +15,7 @@ use ruff_python_semantic::model::SemanticModel;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
 use crate::rules::flake8_simplify::rules::fix_if;
-use crate::settings::options::LineWidth;
+use crate::settings::line_width::LineWidth;
 
 fn compare_expr(expr1: &ComparableExpr, expr2: &ComparableExpr) -> bool {
     expr1.eq(expr2)
